refactor(userspace-dp): relocate reject_reply.rs inline tests to sibling file
reject_reply.rs had grown to 2174 LOC, of which ~1759 lines were an
inline `#[cfg(test)] mod tests { ... }` block wrapping ~380 LOC of cold
reject-reply enqueue production (enqueue_policy_reject_reply /
enqueue_reject_reply). The oversized inline test block buried the
production paths and made the file hard to navigate (codex-review-174 #9).

Extract the test module verbatim into a sibling reject_reply_tests.rs and
wire it via `#[cfg(test)] #[path = "reject_reply_tests.rs"] mod tests;`,
matching the worker_queue / umem / event_stream split convention already
used across the crate. This is pure code motion:

- Production code (reject_reply.rs lines 1-413) is byte-identical to the
  pre-split file; the cold #[cold]/#[inline(never)] enqueue functions are
  untouched.
- Every test body is byte-identical after reversing the 4-space dedent
  out of the former `mod tests {}` wrapper (a re-indent of the new file
  diffs clean against the original inline block).
- `use super::*;` and all imports preserved; tests still resolve under the
  same `afxdp::poll_descriptor::reject_reply::tests::` module path.
- Test count preserved: 25 `#[test]` before == 25 after.

Validation: cargo build clean; `cargo test --release reject_reply` 25/25
ok; full serial suite (`cargo test --release -- --test-threads=1`) 3860
passed / 0 failed / 2 ignored.

Closes #4668
